+++ BalancedSentries/MyPluginInfo.cs
@@ -7,11 +7,11 @@ namespace BalancedSentries
         public const string PLUGIN_GUID = $"{PLUGIN_ORIGINAL_AUTHOR}.{PLUGIN_NAME}";
         public const string PLUGIN_NAME = "BalancedSentries";
         public const string USERS_PLUGIN_NAME = "Balanced Sentries";
-        public const string PLUGIN_VERSION = "1.0.1";
-        public const string PLUGIN_DESCRIPTION = "Adds a 1-6 player version for each ship.";
+        public const string PLUGIN_VERSION = "1.1.0";
+        public const string PLUGIN_DESCRIPTION = "Automatically adjusts B.R.A.I.N. Turret damage and power consumption based on the number of players in the lobby.";
         public const string PLUGIN_ORIGINAL_AUTHOR = "HeX";
         public const string PLUGIN_AUTHORS = "HeX, Airborne";
-        public const string PLUGIN_THUNDERSTORE_ID = "bthnyk/BalancedSentries";
+        public const string PLUGIN_THUNDERSTORE_ID = "ShinyChiefs/BalancedSentries";
     }
 }
 #pragma warning restore CS1591